requirements.txt
flask
flask-cors
openai
langchain
python-dotenv
